+++ packages/nestjs/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @vercel/nestjs
 
+## 0.2.81
+
+### Patch Changes
+
+- @vercel/node@5.8.1
+
 ## 0.2.80
 
 ### Patch Changes